refactor(agent): fold open-session index satellites into session-index
Merge use-open-sessions.ts (useSyncExternalStore bindings) and
session-contracts.ts (ActiveSession + indexOpenByThreadId) into
session-index.ts so one module owns the types, derivation, pub/sub
stores, and React hooks for the sidebar session index. All importers
now pull from @/features/agent/session-index; the module gains
"use client", which is safe because every value importer already
lives in the client graph.
