--- frontend/src/features/agent/session-contracts.ts
@@ -1,16 +1,0 @@
-import type { OpenAgentSession } from "@/features/agent/session-index";
-
-export type ActiveSession = Pick<
-  OpenAgentSession,
-  "projectId" | "cwd" | "paneId" | "id" | "threadId" | "title" | "status" | "focused" | "updatedAt"
->;
-
-export function indexOpenByThreadId(
-  activeSessions: readonly ActiveSession[],
-): Map<string, ActiveSession> {
-  const map = new Map<string, ActiveSession>();
-  for (const session of activeSessions) {
-    if (session.threadId) map.set(session.threadId, session);
-  }
-  return map;
-}
