+++ frontend/src/features/agent/ui/projects-nav-section.tsx
@@ -8,8 +8,11 @@ import {
   useProjectsNavAddProjectEffect,
   useProjectsNavSessionPrefs,
 } from "@/features/agent/ui/projects-nav/use-projects-nav-effects";
-import { useOpenSessions, useSessionActivity } from "@/features/agent/ui/use-open-sessions";
-import { sessionActivity } from "@/features/agent/session-index";
+import {
+  sessionActivity,
+  useOpenSessions,
+  useSessionActivity,
+} from "@/features/agent/session-index";
 import { useProjects } from "@/features/agent/projects/context";
 import { addProjectFromPath, openProjectDirectory } from "@/features/agent/projects/api";
 import { isChatsProject, type Project as ProjectEntry } from "@/features/agent/projects/types";
