--- frontend/src/features/agent/ui/use-open-sessions.ts
@@ -1,18 +1,0 @@
-"use client";
-
-import { useSyncExternalStore } from "react";
-import {
-  getSessionActivity,
-  getOpenSessions,
-  subscribeSessionActivity,
-  subscribeOpenSessions,
-  type OpenAgentSession,
-} from "@/features/agent/session-index";
-
-export function useOpenSessions(): readonly OpenAgentSession[] {
-  return useSyncExternalStore(subscribeOpenSessions, getOpenSessions, getOpenSessions);
-}
-
-export function useSessionActivity() {
-  return useSyncExternalStore(subscribeSessionActivity, getSessionActivity, getSessionActivity);
-}
